Added a refactoring to separate sending file logic from StaticRoute dispatcher. This should allow more flexibility in sending the files from request handlers manually.
