+++ aiohttp/web_urldispatcher.py
@@ -3,7 +3,6 @@ import asyncio
 
 import keyword
 import collections
-import mimetypes
 import re
 import os
 import sys
@@ -19,9 +18,10 @@ from multidict import upstr
 
 from . import hdrs
 from .abc import AbstractRouter, AbstractMatchInfo, AbstractView
+from .helpers import FileSender
 from .protocol import HttpVersion11
 from .web_exceptions import (HTTPMethodNotAllowed, HTTPNotFound,
-                             HTTPNotModified, HTTPExpectationFailed)
+                             HTTPExpectationFailed)
 from .web_reqrep import StreamResponse
 
 
@@ -461,9 +461,6 @@ class StaticRoute(Route):
         self._chunk_size = chunk_size
         self._response_factory = response_factory
 
-        if bool(os.environ.get("AIOHTTP_NOSENDFILE")):
-            self._sendfile = self._sendfile_fallback
-
     def match(self, path):
         if not path.startswith(self._prefix):
             return None
@@ -480,88 +477,6 @@ class StaticRoute(Route):
     def get_info(self):
         return {'directory': self._directory,
                 'prefix': self._prefix}
-
-    def _sendfile_cb(self, fut, out_fd, in_fd, offset, count, loop,
-                     registered):
-        if registered:
-            loop.remove_writer(out_fd)
-        try:
-            n = os.sendfile(out_fd, in_fd, offset, count)
-            if n == 0:  # EOF reached
-                n = count
-        except (BlockingIOError, InterruptedError):
-            n = 0
-        except Exception as exc:
-            fut.set_exception(exc)
-            return
-
-        if n < count:
-            loop.add_writer(out_fd, self._sendfile_cb, fut, out_fd, in_fd,
-                            offset + n, count - n, loop, True)
-        else:
-            fut.set_result(None)
-
-    @asyncio.coroutine
-    def _sendfile_system(self, req, resp, fobj, count):
-        """
-        Write `count` bytes of `fobj` to `resp` starting from `offset` using
-        the ``sendfile`` system call.
-
-        `req` should be a :obj:`aiohttp.web.Request` instance.
-
-        `resp` should be a :obj:`aiohttp.web.StreamResponse` instance.
-
-        `fobj` should be an open file object.
-
-        `offset` should be an integer >= 0.
-
-        `count` should be an integer > 0.
-        """
-        transport = req.transport
-
-        if transport.get_extra_info("sslcontext"):
-            yield from self._sendfile_fallback(req, resp, fobj, count)
-            return
-
-        yield from resp.drain()
-
-        loop = req.app.loop
-        out_fd = transport.get_extra_info("socket").fileno()
-        in_fd = fobj.fileno()
-        fut = asyncio.Future(loop=loop)
-
-        self._sendfile_cb(fut, out_fd, in_fd, 0, count, loop, False)
-
-        yield from fut
-
-    @asyncio.coroutine
-    def _sendfile_fallback(self, req, resp, fobj, count):
-        """
-        Mimic the :meth:`_sendfile_system` method, but without using the
-        ``sendfile`` system call. This should be used on systems that don't
-        support the ``sendfile`` system call.
-
-        To avoid blocking the event loop & to keep memory usage low, `fobj` is
-        transferred in chunks controlled by the `chunk_size` argument to
-        :class:`StaticRoute`.
-        """
-        chunk_size = self._chunk_size
-
-        chunk = fobj.read(chunk_size)
-        while chunk and count > chunk_size:
-            resp.write(chunk)
-            yield from resp.drain()
-            count = count - chunk_size
-            chunk = fobj.read(chunk_size)
-
-        if chunk:
-            resp.write(chunk[:count])
-            yield from resp.drain()
-
-    if hasattr(os, "sendfile"):  # pragma: no cover
-        _sendfile = _sendfile_system
-    else:  # pragma: no cover
-        _sendfile = _sendfile_fallback
 
     @asyncio.coroutine
     def handle(self, request):
@@ -581,36 +496,9 @@ class StaticRoute(Route):
         if not filepath.is_file():
             raise HTTPNotFound()
 
-        st = filepath.stat()
-
-        modsince = request.if_modified_since
-        if modsince is not None and st.st_mtime <= modsince.timestamp():
-            raise HTTPNotModified()
-
-        ct, encoding = mimetypes.guess_type(str(filepath))
-        if not ct:
-            ct = 'application/octet-stream'
-
-        resp = self._response_factory()
-        resp.content_type = ct
-        if encoding:
-            resp.headers[hdrs.CONTENT_ENCODING] = encoding
-        resp.last_modified = st.st_mtime
-
-        file_size = st.st_size
-
-        resp.content_length = file_size
-        resp.set_tcp_cork(True)
-        try:
-            yield from resp.prepare(request)
-
-            with filepath.open('rb') as f:
-                yield from self._sendfile(request, resp, f, file_size)
-
-        finally:
-            resp.set_tcp_nodelay(True)
-
-        return resp
+        file_sender = FileSender(resp_factory=self._response_factory,
+                                 chunk_size=self._chunk_size)
+        return file_sender.send(request, filepath)
 
     def __repr__(self):
         name = "'" + self.name + "' " if self.name is not None else ""
